+++ soroban/argentina_pledge/src/event.rs
@@ -5,27 +5,27 @@ pub(crate) fn set_admin(e: &Env, admin: Address, new_admin: Address) {
     e.events().publish(topics, new_admin);
 }
 
-pub(crate) fn transfer(e: &Env, from: Address, to: Address, id: i128) {
+pub(crate) fn transfer(e: &Env, from: Address, to: Address, id: u64) {
     let topics = (symbol_short!("transfer"), from, to);
     e.events().publish(topics, id);
 }
 
-pub(crate) fn mint(e: &Env, to: Address, id: i128) {
+pub(crate) fn mint(e: &Env, to: Address, id: u64) {
     let topics = (symbol_short!("mint"), to);
     e.events().publish(topics, id);
 }
 
-pub(crate) fn burn(e: &Env, from: Address, id: i128) {
+pub(crate) fn burn(e: &Env, from: Address, id: u64) {
     let topics = (symbol_short!("burn"), from);
     e.events().publish(topics, id);
 }
 
-pub(crate) fn approve(e: &Env, operator: Address, id: i128) {
+pub(crate) fn approve(e: &Env, operator: Address, id: u64) {
     let topics = (symbol_short!("appr"), operator);
     e.events().publish(topics, id);
 }
 
-pub(crate) fn remove_approve(e: &Env, id: i128) {
+pub(crate) fn remove_approve(e: &Env, id: u64) {
     let topics = (symbol_short!("del_appr"),);
     e.events().publish(topics, id);
 }
